Fixed build on windows, which doesn't define rintf
git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/branches/hoomd-0.8@1814 fa922fa7-2fde-0310-acd8-f43f465a7996

+++ src/data_structures/ParticleGroup.h
@@ -94,7 +94,7 @@ class ParticleGroup
 		*/
 		const unsigned int getNumMembers() const
 			{
-			return m_members.size();
+			return (unsigned int)m_members.size();
 			}
 			
 		//! Get a member from the group
